requirements.txt
numpy==1.23.3
pandas==1.5.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.0
plotly==5.10.0
openpyxl==3.0.10
packaging==21.3
requests==2.28.1
python-dotenv==0.21.0
glom==22.1.0
APScheduler==3.9.1.post1
